Use int, not long, for TTL, following the Cassandra APIs. 68 years is a good, long life for a cell

+++ titan-cassandra/src/main/java/com/thinkaurelius/titan/diskstorage/cassandra/astyanax/AstyanaxKeyColumnValueStore.java
@@ -157,7 +157,7 @@ public class AstyanaxKeyColumnValueStore implements KeyColumnValueStore {
                 case TIMESTAMP:
                     return element.getTimestamp();
                 case TTL:
-                    return (long) element.getTtl();
+                    return element.getTtl();
                 default:
                     throw new UnsupportedOperationException("Unsupported meta data: " + meta);
             }
